requirements.txt
rich>=13.7,<14
typer>=0.12,<1
shellingham>=1.5,<2
colorama>=0.4.6,<1
pyperclip>=1.8,<2
loguru>=0.7.3
requests>=2.32.5
beautifulsoup4>=4.14.2
pyyaml>=6.0.0